Don't include external scripts in the DOM if respective keys are not present
Tests in the browser are run in the development environment and there we have Segment and Intercom keys set, but their scripts are not included in tests/index.html. This caused errors because we were calling `analytics` and `Intercom` when not defined.

One solution could have been to include those scripts in the tests/index.html file, but I didn't feel like they belonged to tests, plus that would have introduced duplication.

This solution adds the scripts to the page only if the respective keys are present and checks if `analytics` and `Intercom` are defined before trying to call them, delegating the key presence check to the Brocfile.

If there is no key there should be no scripts.