Adding `onScrollToX` and `onScrollToY` callbacks.
The current `onScroll` method is not sufficient in the
ember-scrollable that supports both horizontal and vertical
scrolling since there is no way of determining what direction
was scrolled without forcing clients to keeping state and detect
the scroll direction (something that is already done by ember-scrollable).
As such, we should deprecate the old usage, and add new callbacks.